Fix #273 use kingston instead kernel
- remove all duplicate file, use @ventureum/kingston instead it
- fix migration, test problem.
- move information of `Kernel` `ACLHandler` `ContractAddressHandler` and
  `Storage` from `ownSolConfig` to `thirdPartySolConfig`.
- pass all(99) unit tests.
