+++ solidity/config/ownSolConfig.js
@@ -60,27 +60,6 @@ export default function (artifacts) {
   Registry.CI = Web3.utils.keccak256('Registry')
 
   /**
-   * Contracts - kernel
-   */
-  class Kernel {}
-  Kernel.Self = artifacts.require('kernel/Kernel')
-  Kernel.RootCI = Web3.utils.keccak256('root')
-
-  /**
-   * Contracts - handlers
-   */
-  // ACLHandler
-  class ACLHandler {}
-  ACLHandler.Self = artifacts.require('handlers/ACLHandler')
-  ACLHandler.CI = Web3.utils.keccak256('ACLHandler')
-
-  // ContractAddressHandler
-  class ContractAddressHandler {}
-  ContractAddressHandler.Self = artifacts.require(
-    'handlers/ContractAddressHandler')
-  ContractAddressHandler.CI = Web3.utils.keccak256('ContractAddressHandler')
-
-  /**
    * Contracts - managers
    */
   // RefundManager
@@ -237,32 +216,6 @@ export default function (artifacts) {
   }
 
   /**
-   * Contracts - storage
-   */
-  class Storage {}
-  Storage.Sig = {
-    SetUint: wweb3.eth.abi.encodeFunctionSignature('setUint(bytes32,uint256)'),
-    GetUnit: wweb3.eth.abi.encodeFunctionSignature('getUint(bytes32)'),
-    SetAddress: wweb3.eth.abi.encodeFunctionSignature(
-      'setAddress(bytes32,address)'),
-    GetAddress: wweb3.eth.abi.encodeFunctionSignature('getAddress(bytes32)'),
-    SetBytes32: wweb3.eth.abi.encodeFunctionSignature(
-      'setBytes32(bytes32,bytes32)'),
-    GetBytes32: wweb3.eth.abi.encodeFunctionSignature('getBytes32(bytes32)'),
-    SetArray: wweb3.eth.abi.encodeFunctionSignature(
-      'setArray(bytes32,bytes32[])'),
-    GetArray: wweb3.eth.abi.encodeFunctionSignature('getArray(bytes32)'),
-    SetUintArray: wweb3.eth.abi.encodeFunctionSignature(
-      'setUintArray(bytes32,uint256[])'),
-    GetUintArray: wweb3.eth.abi.encodeFunctionSignature(
-      'getUintArray(bytes32)'),
-    SetAddressArray: wweb3.eth.abi.encodeFunctionSignature(
-      'setAddressArray(bytes32,address[])'),
-    GetAddressArray: wweb3.eth.abi.encodeFunctionSignature(
-      'getAddressArray(bytes32)')
-  }
-
-  /**
    * Contracts - Modules
    */
   // RegulatingRating
@@ -324,9 +277,6 @@ export default function (artifacts) {
   MockedSale.Self = artifacts.require('./mocks/MockedSale.sol')
 
   return {
-    'Kernel': Kernel,
-    'ACLHandler': ACLHandler,
-    'ContractAddressHandler': ContractAddressHandler,
     'RefundManager': RefundManager,
     'ProjectController': ProjectController,
     'MockedProjectController': MockedProjectController,
@@ -335,7 +285,6 @@ export default function (artifacts) {
     'EtherCollector': EtherCollector,
     'TokenCollector': TokenCollector,
     'TokenSale': TokenSale,
-    'Storage': Storage,
     'MockedSale': MockedSale,
     'RegulatingRating': RegulatingRating,
     'RewardManager': RewardManager,
