+++ solidity/config/thirdPartySolConfig.js
@@ -64,6 +64,55 @@ export default function (artifacts) {
   class Conversion {}
   Conversion.Self = artifacts.require("./Conversion")
 
+  //* -------- Kingston -----------
+  /**
+   * Contracts - kernel
+   */
+  class Kernel {}
+  Kernel.Self = artifacts.require('./Kernel')
+  Kernel.RootCI = Web3.utils.keccak256('root')
+
+  /**
+   * Contracts - handlers
+   */
+  // ACLHandler
+  class ACLHandler {}
+  ACLHandler.Self = artifacts.require('./ACLHandler')
+  ACLHandler.CI = Web3.utils.keccak256('ACLHandler')
+
+  // ContractAddressHandler
+  class ContractAddressHandler {}
+  ContractAddressHandler.Self = artifacts.require(
+    './ContractAddressHandler')
+  ContractAddressHandler.CI = Web3.utils.keccak256('ContractAddressHandler')
+
+  /**
+   * Contracts - storage
+   */
+  class Storage {}
+  Storage.Sig = {
+    SetUint: wweb3.eth.abi.encodeFunctionSignature('setUint(bytes32,uint256)'),
+    GetUnit: wweb3.eth.abi.encodeFunctionSignature('getUint(bytes32)'),
+    SetAddress: wweb3.eth.abi.encodeFunctionSignature(
+      'setAddress(bytes32,address)'),
+    GetAddress: wweb3.eth.abi.encodeFunctionSignature('getAddress(bytes32)'),
+    SetBytes32: wweb3.eth.abi.encodeFunctionSignature(
+      'setBytes32(bytes32,bytes32)'),
+    GetBytes32: wweb3.eth.abi.encodeFunctionSignature('getBytes32(bytes32)'),
+    SetArray: wweb3.eth.abi.encodeFunctionSignature(
+      'setArray(bytes32,bytes32[])'),
+    GetArray: wweb3.eth.abi.encodeFunctionSignature('getArray(bytes32)'),
+    SetUintArray: wweb3.eth.abi.encodeFunctionSignature(
+      'setUintArray(bytes32,uint256[])'),
+    GetUintArray: wweb3.eth.abi.encodeFunctionSignature(
+      'getUintArray(bytes32)'),
+    SetAddressArray: wweb3.eth.abi.encodeFunctionSignature(
+      'setAddressArray(bytes32,address[])'),
+    GetAddressArray: wweb3.eth.abi.encodeFunctionSignature(
+      'getAddressArray(bytes32)')
+  }
+
+
   return {
     'TimeSetter': TimeSetter,
     'Error': Error,
@@ -72,6 +121,11 @@ export default function (artifacts) {
     'ReputationSystem': ReputationSystem,
     'CarbonVoteX': CarbonVoteX,
     'Presale': Presale,
-    'Conversion': Conversion
+    'Conversion': Conversion,
+    'Kernel': Kernel,
+    'ACLHandler': ACLHandler,
+    'ContractAddressHandler': ContractAddressHandler,
+    'Storage': Storage
   }
 }
+
