--- solidity/contracts/handlers/IHandler.sol
@@ -1,5 +1,0 @@
-pragma solidity ^0.4.24;
-
-import '../base/IBase.sol';
-
-contract IHandler is IBase { }
