+++ solidity/contracts/modules/managers/Manager.sol
@@ -1,6 +1,7 @@
 pragma solidity ^0.4.24;
 
-import "../IModule.sol";
+//import "../IModule.sol";
+import "@ventureum/kingston/contracts/modules/IModule.sol";
 import "../Module.sol";
 import "../project_controller/ProjectController.sol";
 import "../milestone_controller/MilestoneController.sol";
