+++ solidity/package.json
@@ -29,6 +29,7 @@
   },
   "devDependencies": {
     "@ventureum/conversion": "^1.0.0",
+    "@ventureum/kingston": "^1.0.0",
     "chai": "^4.1.2",
     "chai-as-promised": "^7.1.1",
     "chai-bignumber": "^2.0.2",
